Move DataColumnSidecarDB to 'db' sub-package
